[SM60 1/2] Upgrade Spidermonkey build system and binaries to 60.9.1
This changes VS to use the same library name as other platforms, since the toolsets are binary-compatible going forward. This makes the build script slightly more streamlined on windows.
Also enable building NSPR on windows as part of the spidermonkey build, instead of compiling it separately.

Tested by: SubitaNeo, Stan

Refs #5859

Differential Revision: https://code.wildfiregames.com/D3115



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24242 3db68df2-c116-0410-a063-a993310a9797