Use C++17 to compile 0 A.D.
Supported compilers are Clang 5, GCC 7, Visual Studio 17, Xcode 9.3

Update Atlas alongside since the replacement APIs are c++17 only.
De-activate the StyledTextCtrl module of WxWidgets on mac, since that is a wrapper around the Scintilla Text editor and doesn't compile with C++17 (as it uses auto_ptr). We don't use the editor, so this is also a win on the compilation time front.

Closes #5862

Differential Revision: https://code.wildfiregames.com/D3166



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24308 3db68df2-c116-0410-a063-a993310a9797