[Petra] do not pass min == max range in moveToRange
Original diff D2512

Reported on forum by gameboy: https://wildfiregames.com/forum/index.php?/topic/27384-strange-landing-on-the-island-and-unable-to-attack/

Units with goal where minrange = maxrange rarely arrive to destination, because they miss it.
See Todo in unitmotion
Relative part:

In the meantime, one should avoid that 'Speed over a turn' > MaxRange - MinRange, in case where min-range is not 0 and max-range is not infinity.

For that reason avoid passing minrange = maxrange from ai.
Also warn in cpp unitmoition when getting this kind of command.

Differential revision: D3149
Reviewed by: @wraitii

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24315 3db68df2-c116-0410-a063-a993310a9797