[SM78 1/2] Upgrade Spidermonkey build system and binaries to 78.6.0
SM78 drops usage of Python 2 in favour of Python 3.

SM78 is the first ESR to actually use rust, which means a new static library is bundled.
This means LLVM-objdump is now a dependency of Spidermonkey, and also 0 A.D.

SM78 also removes the need for NSPR on Windows (Bug 1556646).

Tested by: Stan, Freagarach, mammadori

Refs #5694
Refs #5861

Differential Revision: https://code.wildfiregames.com/D3167



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24332 3db68df2-c116-0410-a063-a993310a9797