Rework bundle creation scripts.
This reworks the OSX bundle script, to run it easily on CI and to further also build the unix tarballs and the windows installer.

Clean up the MacOS bundle detection, by not forcing a compilation-time switch.
Clean up shared library compilation: changing the installation name isn't necessary for 0 A.D.

Differential Revision: https://code.wildfiregames.com/D3161



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24343 3db68df2-c116-0410-a063-a993310a9797