Adjust range visualisation and aura ranges following rP24217
rP24217 / D2759 made ranges account for entity size, making auras larger than they needed to be. This accounts for that.

Units haven't been changed, the delta is not worth modifying imo.

As discussed with Nescio, this means that, for a given aura, larger buildings will now have a larger "aura area". This particularly affects oblong buildings.
A proper solution for this would be to not approximate some structures as a circle, but remains to be implemented.

Note that for current 0 A.D., this effect is unlikely to be a significant gameplay advantage to one civilisation over another.

Reviewed By: Nescio

Differential Revision: https://code.wildfiregames.com/D3169



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24349 3db68df2-c116-0410-a063-a993310a9797