Make units wait a few seconds at waypoints during Patrol.
This makes 'dancing' impossible with patrol, for both single units and formations.
It further makes the formation and regular patrol code more similar.

The diff harcodes an integer wait time as it relies on timer calls instead of range queries, as that was much simpler to implement.
Animations for the waypoint-waiting is also unimplemented.

Reviewed By: Freagarach

Refs #5106

Differential Revision: https://code.wildfiregames.com/D2913



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24360 3db68df2-c116-0410-a063-a993310a9797