Add `getTextSize()` for CButton
Summary: rP22134 introduced the `getTextSize()` to all GUIObjects for getting the rendered size of a text. rP23020 made it only available to CText. `getTextSize()` can also be useful for CButtons so this diff adds the function back to CButtons. There seem to be no other GUIObjects besiders CText and CButton where `getTextSize()` makes sense.

Reviewed By: wraitii
Differential Revision: https://code.wildfiregames.com/D3195



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24378 3db68df2-c116-0410-a063-a993310a9797