[Petra/Ai] Get landaccess of garrison holder if entity is garrisoned
Position of garrisoned entities is undefined, thats why landaccess fails, when asking for position in some cases.
Last replay in ticket is from a23b.

source of the issue was this.target in attackplan got garrisoned so it got undefined possition.

Solution applied is to ask for landaccess of garrisonholder if current entity is garrisoned inside the function itself, instead doing it in every place possible and forget it the future additions. If calling code does not want to deal with garrisoned entities, it should check for that case separately.

For attack plan, when it happened is fine, because it will deal with garrisoned entity later.

Differential revision: D3244
Fixes: #5589

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24436 3db68df2-c116-0410-a063-a993310a9797