[gameplay-a24] Lower cost of briton special tech for grain gather rate
The technology introduced in rP23803 for gauls is a bit overprised.
This is lowering the price to the original cost of rotarymill and lowering gain a bit more because the tech is not just gobal but available in phase earlier. Also it is "almost" half on cost and effect of the current phase II tech which is available for everyone.

Differential Revision: D3253
Reviewed by: borg-

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24460 3db68df2-c116-0410-a063-a993310a9797