Fix Idle formation rearranging when an entity dies.
Idle formations re-arrange whenever a member leaves the formation. Because Idle formation entities can attack nearby units (the formation stays idle for this), any death will re-arrange the formation, causing the entities to move in formation before going back to attack.

This looks odd and makes formation less usable. Given that it's tricky to change the controller state with the current code, this instead reforms the formation on a timer, only when all members are IDLE.

It fixes the issue and looks generally similar in other cases.

Reported By: wowgetoffyourcellphone

Reviewed By: Freagarach

Differential Revision: https://code.wildfiregames.com/D3236



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24467 3db68df2-c116-0410-a063-a993310a9797