Allow picking a default formation for walk (and walk-like) orders.
This allows choosing a "default formation", which is activated automatically for units given walk orders (and attack-walk etc.).
Conversely, units in formation that are given a gather/build/... order are taken out of formation and given the order individually.
The default formation can be selected by right-clicking on any formation icon.

This leverages formations for walking, where they are quite efficient (in fact, perhaps too efficient), while circumventing issues with various orders.

Choosing the "null formation" as the default formation de-activates the behaviour entirely, and plays out exactly like SVN.

This makes it possible to queue a formation-order then a noformation-order (i.e. walk then repair), though the behaviour isn't very flexible.

For modders, it should be relatively easy to change the setup for each order, and/or to force deactivating/activating formations in general.

Tested by: Freagarach, Angen

Refs #3479, #1791.
Makes #3478 mostly invalid.

Differential Revision: https://code.wildfiregames.com/D2764



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24480 3db68df2-c116-0410-a063-a993310a9797