Regroup special technologies.
Some techs are shared between multiple civs and having them in a folder that is named differently (successors/hellenes) can be confusing.
This patch deprecates the subfolders and moves all techs in the main folder.

(Also did svn propset on *.json to not have to do it for every changed template.)

Patch by: @Nescio
Differential revision: D888
Comment by: @fatherbushido



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24558 3db68df2-c116-0410-a063-a993310a9797