Use correct hotkey in tooltip of Reference Suite's Close Buttons
The various pages of the Reference Suite (`structree`, `civinfo`, `viewer`) all
use the same `CloseButton` object, and the same hotkey to close each page.

The hotkey that is listened for is the one assigned to `cancel`, however the
tooltip was attempting to display the hotkey assigned to `close` (which doesn't
actually exist).



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24601 3db68df2-c116-0410-a063-a993310a9797