Fix PetraAI not building many houses.
Previous to r24394 / rP24394 the AI got its population bonus per entity from the Cost component, which every entity happened to have.
Thereafter, the population bonus was handled by a separate component, which was not always present leading to `(void 0)` additions to the pop cap and PetraAI not constructing houses.

Differential revision: D3360
Reviewed by: @Angen, @wraitii



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24613 3db68df2-c116-0410-a063-a993310a9797