[Gameplay A24] - Add Carnyx unit to Gaul.
The carnyx was a long, bronze, Celtic trumpet and used on the battlefield; its sound would intimidate opponents.
This patch introduces a trainable infantry champion for the carnyx-using actor, at the assembly, and adds a generic template and an aura file, which lowers the attack damage of nearby enemy soldiers a bit.
While it has an aura that could make a difference, it has no training restrictions (unlike heroes); to compensate for that, it has a higher cost, less resistance and less attack damage than other champions. If it turns out to make the Gauls too strong, its values can be adjusted in the future.
Having this (historically attested) unit would make the Gauls less similar to other civs.

The specific name was suggested by @genava55 on the forums.
Icon adapted by @Stan from @niektb.

Original patch by: @Stan
Redone by: @Nescio
Differential revision: D3320
Reviewed by: @borg-
Comments by: @Freagarach



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24634 3db68df2-c116-0410-a063-a993310a9797