Fix lag on OOS with AIs
Following rP24462, the OOS debugSerializer would try to serialize AI objects entirely. This is extremely slow (might be an infinite loop in some cases).
Instead, it should print the result of Serialize() in those cases.

(The original idea was that it would print more debugging information, but in practice it seems to mostly print things that show up when diffing but aren't actually sources of OOS, so essentially garbage).

Fixes #5917

Differential Revision: https://code.wildfiregames.com/D3361



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24641 3db68df2-c116-0410-a063-a993310a9797