Fix hotkeys not releasing properly.
rP24215 changed hotkeys to use scancode, but didn't change g_keys to reflect that, and this could break hotkey release.
This fixes that by explicitly using scancodes. Note that we might want a g_keys map in the future, but it seems un-necessary at the moment.

Also remove the last remnants of 'negated' hotkeys, which were disabled following rP15980.

Tested By: OptimusShepard

Fixes #5922

Differential Revision: https://code.wildfiregames.com/D3379



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24645 3db68df2-c116-0410-a063-a993310a9797