Format hotkey tooltip when an alternate key-name is used
Some keys have multiple names that they may be identified by in config files,
such as the Escape key which may be identified by either "Esc" or "Escape", or
the number-pad keys which may be identified with a name in the form "Num{xxx}"
(e.g. "NumMinus") or the alternate "Keypad {x}" (e.g. "Keypad -").

The alternate names were not being picked up and formatted correctly when
displayed in tooltips. Now they are.


Accepted By: wraitii
Differential Revision: https://code.wildfiregames.com/D3358



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24668 3db68df2-c116-0410-a063-a993310a9797