Use UnitMotion to predict target position in Attack.js to prevent 'dancing'
Attack.js can use UnitMotion to calculate the position of the unit in the future, accounting for odd movements such
as zigzags, turnarouds and early stops (to the extent of the current order).
This improves the resilience of units against the 'dancing' trick.

The linear interpolation is kept as a failsafe and to avoid an edge case in the new prediction code.

Patch by: bb

Refs #5106

Differential Revision: https://code.wildfiregames.com/D3225



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24701 3db68df2-c116-0410-a063-a993310a9797