Don't pause music when pausing the game / Fix victory music in SP.
Victory music doesn't work in A24 SP. That's because the modal pauses the game, and pausing the game pauses music.
This has been the case since rP13347.

Auto-pausing the game (structree, ...) didn't really happen in the past, but A24 makes it relatively common, and I think pausing the music is a bit jarring then, so this simply removes that.

Reported by: Langbart

Fixes #5941

Differential Revision: https://code.wildfiregames.com/D3433



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24744 3db68df2-c116-0410-a063-a993310a9797