Fix target height computation when launching projectiles.
The Y coordinate at which to fire a projectile is currently assumed to be the target's current Y, which is incorrect if the target is moving on a slope.
This fixes that.

Note that this was purely visual, since projectiles still hit the target regardless, as the height component is totally ignored, even if the projectile is underground (in fact, the projectile's position is not known in DelayedDamage::MissileHit, which just assumes it lands where it said it would when fired).

As noted by bb in rP24701

Fixes #5939

Differential Revision: https://code.wildfiregames.com/D3425



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24766 3db68df2-c116-0410-a063-a993310a9797