Get the public IP from the lobby if not using STUN.
Follow-up to rP24728. This makes it again possible to host without STUN via the lobby.
The lobby bot will answer the host "Register" command with the external IP. This is only sent to the host, avoiding IP leakage.

There is a small window in which a client might try to join and the public IP isn't up, and the request goes through, but that seems rather unlikely to be a problem in practice.

Refs #5913

Differential Revision: https://code.wildfiregames.com/D3490



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24858 3db68df2-c116-0410-a063-a993310a9797