Allow capture when falling back to "WalkAndFight" in COMBAT.APPROACHING
COMBAT.APPROACHING will fallback to WalkAndFight if the order is forced and the movement fails (happens e.g. on target death).
That behaviour was added in rP22425 (and tweaked in rP22567), to make the unit move to the general position asked by the player as a fallback.
However, it did not allow capture, which meant units would occasionally start attacking structures, instead of capturing (the default).
This makes that fallback allow capture in all cases, which is generally more expected than the reverse.

Fixes #6003

Differential Revision: https://code.wildfiregames.com/D3548



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24866 3db68df2-c116-0410-a063-a993310a9797