Fix no-formation orders after several formation orders.
When giving several queued formation-walk orders, then a queued non-formation order (such as gathering), the gather order would be silently dropped. This did not happen if only one formation-walk order was given, which is why this issue was not noticed in rP24480

To fix it, explicitly only drop pre-"leave formation" orders in the queue for formation members.

Reported By: faction02

Differential Revision: https://code.wildfiregames.com/D3550



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24869 3db68df2-c116-0410-a063-a993310a9797