[Gameplay] - Search for resources near a current location before the init-pos.
This should bring the behaviour back to pre-A24 and means entities will stray a lot less when gathering.
The entity will search close to the current position, which is mostly next to a dropsite.
When there is nothing found there, the entity searches nigh the initPos.

Differential revision: D3607
Comments by: @wraitii
Tested by: @Nescio



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25000 3db68df2-c116-0410-a063-a993310a9797