Convert InitAttributes disabled techs/templates & starting tech to triggers.
Disabled technologies & templates are not really gamesetting concerns, since they are map-specific and de-activating these generally does not make sense.
Starting technologies is unused and not the correct granularity level for game setup.

Triggers do a better job of specifying these things on a per-map basis (in fact, some maps already use triggers for that).

Changes:
- StartingTechnologies was used by the cart sandbox and is simply removed.
- DisabledTechnologies was used by Polar Sea RM, which already had triggers -> into triggers.
- DisabledTemplates was used for the Ptol Lighthouse on a few maps -> since we changed the ability to "a lot of range", it's no longer OP on these maps and the restriction is simply lifted.
- DisabledTemplates was used on Death Canyon Invasion Force -> into triggers.

Differential Revision: https://code.wildfiregames.com/D3712



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25088 3db68df2-c116-0410-a063-a993310a9797