Color incompatible mods and add filter
This helps to see compatible mods in the list, which can get a bit long.
For incompatible mods in disabled section use gray colour.
For incompatible mods in enabled section use red colour (what can happen when using arguments to start the game or they can be in config file).
Add option to filter compatible mods, this does not affect enabled section, because if there is incompatible mod, it is error and it needs to be always visible.
Add function recomputeCompatibility and g_ModsCompatibility to avoid recomputing aredependenciesmet when coloring mods, because every column is colored separately.
Disable enable mod button if mod is incompatible.
Remove all dependencies met messages as they are now replaced by this mechanism.

fix getSelectedModUrl in process

Differential revision: D3687
Fixes: #6053
Tested by: @Freagarach
Comments by: @vladislavbelov

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25095 3db68df2-c116-0410-a063-a993310a9797