Optimise FindWalkAndFightTargets.
FindWalkAndFightTargets is used during attack-walk (and a few other situations) to find new entities to attack. This function can be a bit slow, taking large chunks of time during battles.

This optimises it by assuming that one of the surrounding unit will match preferred criteria (which, for most soldiers, are 'Human'), thus returning the first attackable entity. In the worst case, it should still be slightly faster than the current code.

Differential Revision: https://code.wildfiregames.com/D3446



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25102 3db68df2-c116-0410-a063-a993310a9797