Always pass a civcode to Viewer when instatiated by the session gui
Previously this was only done for technologies.

The Viewer tries to determine an appropriate civcode from a passed template,
however trainable gaia-animals have a template-assigned civ of `gaia` (and
technologies don't have one at all).

The behaviour of the Viewer isn't changed by this commit: it will still prefer
whatever civ is set within a template before falling back to the passed civcode.

However, trained gaia-animals now have something to fall back to, permitting the
appropriate build list (the list of buildings that train the animal) to be
displayed.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25110 3db68df2-c116-0410-a063-a993310a9797