Remember OOS on a per-client basis.
Change the OOS notification logic to remember the OOS-ness of each client. Reset it on client leave.
The server will thus continue checking for OOS if the OOS-client leaves.
This is convenient to ignore observer OOS, or wait for an OOS player without restarting the game.

Also add the turn number to the OOS dump, to fix #3348: particularly following rP25001 the turn is likely to not be the same between different clients.

Agree by: asterix

Differential Revision: https://code.wildfiregames.com/D3753



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25170 3db68df2-c116-0410-a063-a993310a9797