Further hotkey work: clear hotkeys when an input box/the console gets focus.
This is necessary following rP25180, which now handles key-up in input to prevent hotkeys from firing then.
Hotkeys are a global system, which means opening e.g. the chat window or the summary does't reset things. If you press "S" to scroll the in-game camera down, then press 'return' to open the chat, the camera keeps scrolling down. Following rP25180, it never released since it never got the keyup.

To handle this nicely, we need to explicitly clear hotkeys when input boxes get focus (respectively the console).

Differential Revision: https://code.wildfiregames.com/D3797



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25186 3db68df2-c116-0410-a063-a993310a9797