Add Upkeep component.
Adds a new component effectively handling negative resource trickles.
Prevents resources from going negative and provides an effect for not being able to pay.
The effect chosen for 0 A.D. is having the entity not being controllable.

Not used in the public mod itself, but it is in quite some mods.

Based on an idea of @Nescio
Differential revision: D1323
Comments by: @Angen, (@smiley,) @Stan



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25191 3db68df2-c116-0410-a063-a993310a9797