ParamNode: switch to utf8, ToXML -> ToXMLString, test tweak, PCH tweak
This changes ParamNode to use UTF8 values internally (XMB files are UTF8 since rP15309).
This removes the need for a lot of conversions, speeding things up and allows cleaning up the validator interface & a few other callsites.
ConstructJSVal could be a tad slower because of UTF8->16 conversions within Spidermonkey; but the difference is unlikely to be noticeable in practica.

Also:
- Changes `ToXML` to `ToXMLString` for clarity.
- Add a simple "op" test & show a particular behaviour of merge nodes that I intend to change somewhat in D3830.
- Remove Component.h from simulation2 PCH - brought in too much.

Tested by: langbart

Differential Revision: https://code.wildfiregames.com/D3834



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25228 3db68df2-c116-0410-a063-a993310a9797